Enabled data generation and benchmark runs.
WARNING: THIS CHANGE REQUIRES FIXES IN EVERY generate.c FILE!
We need to specify the output file in both the top-level harness and each
individual generate file. The harness is a one-time fix, but the generate file
will have to be changed in every kernel and algorithm.

Currently, everything but the metric collection and reporting works in scons.